Decompose form remote submit procedure
NOTE: File uploading support is removed in this commit, since we will
employ a new method (FormData) to handle ajax sooner.
